backend/requirements.txt
fastapi
uvicorn[standard]
beanie
motor
google-generativeai
PyGithub
celery
redis
qdrant-client
python-jose[cryptography]
cryptography
passlib[bcrypt]
python-dotenv
httpx
pydantic
pydantic-settings
pytest
pytest-asyncio
black
langgraph
langchain
email-validator